requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
